Refactor command package structure
